+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Bumped `tokio-tungstenite` to version 0.20.0 that is not vulnerable to
   https://rustsec.org/advisories/RUSTSEC-2023-0052 via the `webpki` crate.
 - Added support for addresses without any transactions.
+- Bumped the minimum supported Blockbook version to [commit `95ee9b5b`](https://github.com/trezor/blockbook/commit/95ee9b5b).
 
 ## 0.1.0